Append ProductViewHelper methods to ProductHelper
The original Spree::ProductHelper is a better home for these methods as
they affect product views.
